Break the plugin into a separate file
